+++ aquadoggo/src/db/stores/test_utils.rs
@@ -3,13 +3,13 @@
 use std::convert::TryFrom;
 use std::str::FromStr;
 
-use p2panda_rs::document::DocumentId;
+use p2panda_rs::document::{DocumentId, DocumentView, DocumentViewFields, DocumentViewId};
 use p2panda_rs::entry::{sign_and_encode, Entry};
 use p2panda_rs::hash::Hash;
 use p2panda_rs::identity::{Author, KeyPair};
 use p2panda_rs::operation::{
-    Operation, OperationEncoded, OperationFields, OperationId, OperationValue, PinnedRelation,
-    PinnedRelationList, Relation, RelationList,
+    AsOperation, Operation, OperationEncoded, OperationFields, OperationId, OperationValue,
+    PinnedRelation, PinnedRelationList, Relation, RelationList,
 };
 use p2panda_rs::schema::SchemaId;
 use p2panda_rs::storage_provider::traits::{
@@ -19,6 +19,7 @@ use p2panda_rs::storage_provider::traits::{EntryStore, OperationStore};
 use p2panda_rs::test_utils::constants::{DEFAULT_PRIVATE_KEY, TEST_SCHEMA_ID};
 
 use crate::db::provider::SqlStorage;
+use crate::db::traits::DocumentStore;
 use crate::graphql::client::{EntryArgsRequest, PublishEntryRequest};
 use crate::test_helpers::initialize_db;
 
@@ -162,6 +163,45 @@ pub fn test_key_pairs(no_of_authors: usize) -> Vec<KeyPair> {
     }
 
     key_pairs
+}
+
+pub async fn insert_entry_operation_and_view(
+    provider: &SqlStorage,
+    key_pair: &KeyPair,
+    schema: &SchemaId,
+    document_id: Option<&DocumentId>,
+    operation: &Operation,
+) -> (DocumentId, DocumentViewId) {
+    let request = construct_publish_entry_request(provider, operation, key_pair, document_id).await;
+
+    let operation_id: OperationId = request.entry_encoded.hash().into();
+    let document_id: DocumentId = request.entry_encoded.hash().into();
+    let document_view_id: DocumentViewId = request.entry_encoded.hash().into();
+
+    let author = Author::try_from(key_pair.public_key().to_owned()).unwrap();
+
+    provider.publish_entry(&request).await.unwrap();
+    provider
+        .insert_operation(&OperationStorage::new(
+            &author,
+            operation,
+            &operation_id,
+            &document_id,
+        ))
+        .await
+        .unwrap();
+
+    let document_view_fields =
+        DocumentViewFields::new_from_operation_fields(&operation_id, &operation.fields().unwrap());
+    provider
+        .insert_document_view(
+            &DocumentView::new(&document_view_id, &document_view_fields),
+            schema,
+        )
+        .await
+        .unwrap();
+
+    (document_id, document_view_id)
 }
 
 /// Construct and return a storage provider instance backed by a pre-polpulated database. Passed parameters
